Splits skill children between skills and activities
