--- src/app/modules/item/components/item-children-edit/item-children-edit.component.scss
@@ -1,48 +1,0 @@
-@import "src/colors.scss";
-
-.section-chapter-wrapper {
-  padding: 0 2rem;
-}
-
-.add-item-container {
-  padding-left: 5.8333rem;
-  padding-right: 1.6667rem;
-}
-
-.reorder-handle {
-  cursor: grab;
-}
-
-.main {
-  flex: 1;
-}
-
-.type {
-  color: var(--base-color);
-}
-
-.spacer {
-  width: 2.9rem;
-}
-
-.score-weight-container {
-  display: flex;
-  justify-content: flex-end;
-}
-
-.score-weight {
-  display: flex;
-  align-items: center;
-  margin-bottom: 1rem;
-}
-
-.score-weight-label {
-  margin-right: 1rem;
-}
-
-.propagation-state-wrapper {
-  width: 3rem;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
